Float Formatting

+++ CS126L/1.4/bmicalculator.c
@@ -3,7 +3,7 @@
 
 
 void main() {
-    int height, weight, calculatedBMI; // Declartions for userinput and the calculated input off of that
+    float height, weight, calculatedBMI; // Declartions for userinput and the calculated input off of that
 
    printf("BMI Values:\n\n" // Display calculation chart
 "Underweight: less than 18.5\n"
@@ -12,10 +12,10 @@ void main() {
 "Obese: 30 or greater\n\n");
 
     printf("Enter your weight(in pounds): "); // Get user input for height and weight
-    scanf("%d", &weight);
+    scanf("%f", &weight);
     printf("Enter your height(in inches): ");
-    scanf("%d", &height);
+    scanf("%f", &height);
 
     calculatedBMI = (weight * 703) / (height * height); // BMI calculation
-    printf("Your calculated Body Mass Index (BMI) is: %d\n\n", calculatedBMI);
+    printf("Your calculated Body Mass Index (BMI) is: %.2f\n\n", calculatedBMI);
 }
